refactor for resolving cyclic dependencies
