+++ src/databricks/labs/ucx/source_code/linters/spark_connect.py
@@ -11,7 +11,7 @@ from databricks.labs.ucx.source_code.base import (
 )
 from databricks.sdk.service.compute import DataSecurityMode
 
-from databricks.labs.ucx.source_code.linters.python_ast import Tree, TreeHelper
+from databricks.labs.ucx.source_code.python.python_ast import Tree, TreeHelper
 
 
 @dataclass
