+++ tests/unit/source_code/samples/dfsa/spark_read_format_load.py
@@ -1,0 +1,2 @@
+# Databricks notebook source
+spark.read.format("delta").load("s3a://prefix/some_file.csv")
