+++ tests/unit/source_code/test_dfsa.py
@@ -16,21 +16,19 @@ def test_dfsa_does_not_collect_erroneously(simple_dependency_resolver, migration
 
 
 @pytest.mark.parametrize(
-    "source_path, dfsa_paths, supported",
+    "source_path, dfsa_paths",
     [
-        ("dfsa/create_cloud_files.sql", ["s3a://db-gtm-industry-solutions/data/CME/telco/PCMD"], False),
-        ("dfsa/create_location.sql", ["s3a://db-gtm-industry-solutions/data/fsi/capm/sp_500/"], True),
-        ("dfsa/select_read_files.sql", ["s3a://db-gtm-industry-solutions/data/fsi/capm/sp_500/file.csv"], True),
-        ("dfsa/select_format.sql", ["hdfs://examples/src/main/resources/users.parquet"], True),
+        ("dfsa/create_location.sql", ["s3a://db-gtm-industry-solutions/data/fsi/capm/sp_500/"]),
+        ("dfsa/select_read_files.sql", ["s3a://db-gtm-industry-solutions/data/fsi/capm/sp_500/file.csv"]),
+        ("dfsa/select_format.sql", ["hdfs://examples/src/main/resources/users.parquet"]),
+        ("dfsa/spark_read_format_load.py", ["s3a://prefix/some_file.csv"]),
     ],
 )
 def test_dfsa_collects_sql_dfsas(
-    source_path, dfsa_paths, supported, simple_dependency_resolver, migration_index, mock_path_lookup
+    source_path, dfsa_paths, simple_dependency_resolver, migration_index, mock_path_lookup
 ):
-    if (
-        not supported
-    ):  # SQL expression not supported by sqlglot for Databricks, keeping the test data for once we drop sqlglot
-        return
+    """SQL expression not supported by sqlglot for Databricks, restore the test data below for once we drop sqlglot
+    ("dfsa/create_cloud_files.sql", ["s3a://db-gtm-industry-solutions/data/CME/telco/PCMD"]),"""
     maybe = simple_dependency_resolver.build_notebook_dependency_graph(Path(source_path), CurrentSessionState())
     assert not maybe.problems
     context = LinterContext(migration_index, CurrentSessionState())
